// src/interfaces/data.ts
export type Idata = {
  menopausal_status: number | null;
  platelets: number | null;
  albumin: number | null;
  globulin: number | null;
  diameter_of_ovary: number | null;
  solid: number | null;
  papillary_projection: number | null;
  cyst_wall: number | null;
  septum: number | null;
  ascites: number | null;
  ca_125: number | null;
};